Fallback to getMethod if getField didn't work
Also, removed unnecessary duplicated crossScalaVersions
